[DF-9778] warnings fixed in Sources/Tools
Former-commit-id: 9236991709fc44c99335a9e429a0b0b7c4237b83

+++ Sources/Tools/CommandLine/CommandLineParser.cpp
@@ -158,7 +158,7 @@ DAVA::int32 CommandLineParser::GetCommandPosition(const DAVA::String& command)
     int32 position = INVALID_POSITION;
 
     const Vector<String>& commandLine = Core::Instance()->GetCommandLine();
-    for (int32 i = 0; i < (int32)commandLine.size(); ++i)
+    for (size_t i = 0; i < commandLine.size(); ++i)
     {
         if (command == commandLine[i])
         {
@@ -193,6 +193,6 @@ String CommandLineParser::GetCommandParamAdditional(const String& command, const
 int32 CommandLineParser::GetCommandsCount()
 {
     const Vector<String>& commandLine = Core::Instance()->GetCommandLine();
-    return (int32)commandLine.size();
+    return static_cast<int32>(commandLine.size());
 }
 };
